actually fix build issues
