+++ app/post/page.tsx
@@ -17,15 +17,9 @@ import {
 import { Input } from "../_components/ui/input"
 import FormButton from "../_components/ui/formButton"
 import { Textarea } from "../_components/ui/textarea"
+import { formSchema } from '../_components/schemas'
 
-export const formSchema = z.object({
-    title: z.string().min(2, {
-        message: 'Title must be at least 2 characters long'
-    }),
-    magnet: z.string(),
-    description: z.string(),
 
-})
 
 export default function PostAModel() {
